Stub for moving units
Adjust the pointerdown handler code to be able to tell the difference between selecting a unit and clicking on open ground. If the player clicks on open ground with units selected, that is an instruction to move the selected units there. However that instruction means messaging GameServer, which isn't done yet.