Server-side ticking and basic movement
Implement ticking to step the game in GameServer. This includes some slightly tricky code to make sure ticks are scheduled at the right rate, given timers can run late. Units commanded to move now move in a straight line to their target position on GameServer - but this isn't synced to GameClient at the moment, so this can't be seen yet.