Basic game state update message
Implement a basic binary game state message that the GameServer sends to the client every tick. The client then reads this message and updates units accordingly. Currently this just sends the full state for all units in the game. This will do for now, but a smarter strategy will be implemented in future. The movement is a bit janky due to the server tick rate, but it will probably get smoothed out once there's code added to smoothly handle irregular updates from the network.