Implement multiplayer host and peer game modes
With the host and peer game modes, now a multiplayer game can be started with each player controlling their own set of units! One tricky part of this is ensuring the host and peer are synchronised and ready to start at the same time. This is done by having the peer post ready messages to the host regularly, and once the host responds with a "start" message, both proceed to start the game.