Pass Construct object data from client to server
GameServer needs details of objects like their origin, image points and collision polygon. These are stored in the Construct objects on the client side. The host now collects this data from Construct objects and posts it to GameServer so it can also use it for things like correctly positioning turrets on the unit platform.