Projectiles (part 2 of 2)
GameClient now receives network event messages, and handles projectiles being fired. It creates a ClientProjectile and just advances it at its speed and angle. It doesn't need to do anything more than that, as projectile movement is entirely predictable so no further communication is necessary, and anyway they only exist to show the server game state to the player.