Collision detection (part 1 of 3)
Rework NetworkEvent in to a base class, and two derived classes: one for the existing FireProjectile event, and a new one for ProjectileHit, representing a projectile hitting a target. This just sends the projectile ID, so the client can destroy the corresponding projectile, and the position, so the client can show an effect happening there (e.g. create an explosion).