Collision detection (part 2 of 3)
Implement a CollisionShape class, which rotates the collision polygon of an object and can check if a point is inside the shape. Add collision detection for projectiles as they move (currently using a basic brute force approach), and when they hit something, destroy them and send a ProjectileHit network event.