Fix range calculation bug
Fix a small bug with range calculation. Turrets must measure their range from their own platform position to the target's platform position for fairness. Previously turrets measured range from the turret position, but turrets are slightly offset from the platform position, which could make it possible for a turret to be able to fire at a target that could not fire back.