Distinguish taps and drags in SelectionManager
SelectionManager now saves the start position of every pointerdown event. If a pointer moves a certain distance, it gets marked as a dragging pointer instead. If it doesn't move that far by the pointerup event, it's treated as a tap, and performs things like unit selection and movement. This allows SelectionManager to distinguish taps from drags, which is an important precursor for adding dragging features like a drag selection box.