Implement pinch-to-zoom gesture
This allows using two touch pointers to both scroll the view (based on their mid-point) and zoom (based on the change in distance between them). The code is fairly tricky to get right but fortunately can re-use some of the code for mouse panning and wheel zoom. The pinch-to-zoom gesture can be a bit wobbly when moving so probably needs some smoothing, and could do with a momentum effect on scrolling too.