Improve turret targeting algorithm
When multiple targets are available in range, aim for the one that requires the least turret rotation to aim at. This reduces wasted turret rotation time in combat, and avoids the tendency for all turrets to end up aiming left (because the order of iteration of units tends to go left to right in the current setup).