Add PingManager
Add a client-side class to manage sending pings to the server. The server responds to pings with "pongs" with the server time, allowing clients to synchronise clocks. This also involves a way to send messages from GameServer to a specific player. PingManager currently takes a measurement, but doesn't do anything with it yet.