Client unit interpolation (part 3 of 3)
When receiving updates from the network, use the server time in the message to insert values to unit's timelines. Then every tick units update according to the current values in the timelines at the current simulation time (i.e. a delayed time behind the server time). This allows for things like interpolating between the previous and next values.