Client network event buffering
Use a stepped timeline to schedule network events for their intended simulation time. Adjust the handling of late position updates to be handled by ClientPlatform, as we still want to handle late network events from a stepped timeline. Late events can still be handled, for example by ticking a projectile that was fired by the time it was late, so it jumps ahead to where it is meant to be.