Optimize unit ticking
Only tick units that need it. Keep a set of units to tick. Units start of ticking when they're created, but if they are static and nothing is in the timelines then they can opt-out of ticking to save client CPU time. If any network updates arrive, the unit will start ticking again.