Lazy-create client units from full updates
Rather than sending all unit data in the "create-initial-state" message, instead just let the client create them based on full update messages. To make this possible, add a byte to the full update with the unit player number. Add a fade behavior to units so they fade in when first created.