Optimize timeline insertion
Most of the time, new values added to a timeline are new and so are appended to the end. Therefore it's better to iterate the timeline backwards in Add(), since it's likely it will insert the value on the first iteration, as opposed to the last. This is particularly important with the position history timeline which keeps lots of entries and always inserts to the end.