Fix freezing after suspend
If the client is suspended its calculated server time difference will become wrong by the time it was suspended for, and end up so far behind it just freezes. To work around this make PingManager just catch up to the current server time difference if it is wrong by more than 1 second.