Add pathfinding fundamentals
A PathfindingController object is now placed in the game layout with the Pathfinding behavior. A new PathfindingController class on the client will be responsible for pathfinding, using the behavior API. A new debug layer with a debug tilemap is added to display the pathfinding map state. The client-side message handler has also been renamed for consistency with the server.