Implement basic pathfinding waypoints
Add support for async messaging from server to client. Use this to get the host client to perform pathfinding on behalf of the server. This allows finding paths around obstacles as a series of waypoints. Currently units move from one waypoint to the next using an inefficient stop-start system, but this will be improved upon. Unit movement has also been refactored to clarify the separate phases.