Smoothed movement along waypoints
This improves the movement algorithm to take in to account the unit's turn circle while approaching a waypoint, and allows it to turn early such that it will still be on the right line after finishing its turn. There is an additional rule that slows down the unit if its turn circle would go more than half way to the following waypoint, since that makes an overshoot likely.