Draw terrain on minimap
Add a third minimap layer for terrain, which only redraws on startup or when the window is resized. This generates pixel data based on the pathfinding obstacle map and displays it in the canvas. Also add a Tiled Background for the minimap background (otherwise the background color would have to be in the terrain canvas).