Refactor unit movement to UnitMovementController
Previously all the complex logic for moving units along paths was in the MovableUnitPlatform class. The logic is going to get even more complicated with unit avoidance. So move all the path-following logic to a separate UnitMovementController class that is exclusively concerned with managing the movement of an individual unit. This leaves a much simpler MovableUnitPlatform class and is a better architecture for further extending UnitMovementController logic.