Add debug state to unit data
For developing things with complex internal states like managing movement state, it's useful for the client to be able to show the current state. To allow this add a new debug state property for units on the server - currently just a uint8 - and send it to clients in full and delta updates. The client can then show this in a text object over the unit. This is enabled with a couple of flags so it can easily be turned off again in future and save the bandwidth overhead.