Add separate obstacle collision polygon
Rename the existing unit platform collision polygon to the "full" collision polygon. Add a new ObstacleCollision animation to the unit platform and pass it along to the server as well. This allows using two separate collision masks on the server: the full collision mask is used for detecting projectile impacts, and the obstacle collision mask is used for detecting collisions with other units (and in future the scenery). The obstacle collision mask is currently set to a diamond shape, which allows more flexibility with rotating close to other units, at the expense of visually partially overlapping in some circumstances.